Move output as package and testing-library
