+++ packages/design-system/package.json
@@ -42,6 +42,7 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@googleforcreators/testing-library": "*",
     "@storybook/addon-actions": "^6.3.0",
     "@storybook/addon-knobs": "^6.3.0",
     "@testing-library/react": "^12.1.2",
