+++ packages/output/package.json
@@ -1,0 +1,37 @@
+{
+    "name": "@googleforcreators/output",
+    "private": true,
+    "engines": {
+      "node": ">= 16",
+      "npm": ">= 7.3"
+    },
+    "exports": {
+      "default": "./src/index.js"
+    },
+    "main": "./src/index.js",
+    "type": "module",
+    "dependencies": {
+      "@googleforcreators/animation": "*",
+      "@googleforcreators/design-system": "*",
+      "@googleforcreators/elements": "*",
+      "@googleforcreators/fonts": "*",
+      "@googleforcreators/i18n": "*",
+      "@googleforcreators/masks": "*",
+      "@googleforcreators/patterns": "*",
+      "@googleforcreators/react": "*",
+      "@googleforcreators/rich-text": "*",
+      "@googleforcreators/story-editor": "*",
+      "@googleforcreators/units": "*",
+      "draft-js": "^0.11.7",
+      "draft-js-export-html": "^1.4.1",
+      "draft-js-import-html": "^1.4.1",
+      "draftjs-filters": "^2.5.0",
+      "flagged": "^2.0.1",
+      "styled-components": "^5.3.3"
+    },
+    "devDependencies": {
+      "@googleforcreators/testing-library": "*",
+      "@testing-library/react": "^12.1.2",
+      "prop-types": "^15.8.0"
+    }
+}
