+++ packages/output/src/story.js
