+++ packages/output/src/utils/ampBoilerplate.js
