+++ packages/output/src/utils/getLongestMediaElement.js
