+++ packages/output/src/utils/getPreloadResources.js
