+++ packages/output/src/utils/getUsedAmpExtensions.js
