+++ packages/output/src/utils/test/getUsedAmpExtensions.js
