+++ packages/story-editor/src/app/story/utils/test/getStoryPropsToSave.js
@@ -14,10 +14,13 @@
  * limitations under the License.
  */
 /**
+ * External dependencies
+ */
+import { getStoryMarkup } from '@googleforcreators/output';
+/**
  * Internal dependencies
  */
 import getStoryPropsToSave from '../getStoryPropsToSave';
-import getStoryMarkup from '../../../../output/utils/getStoryMarkup';
 
 jest.mock('../../../../output/utils/getStoryMarkup', () => jest.fn());
