+++ packages/story-editor/src/index.js
@@ -36,8 +36,10 @@ export * from './components/panels/shared';
 export * from './components/richText';
 export * from './components/checklist';
 export * from './components/checklistCard';
+export * from './components/elementLink';
 export { useVideoTrim, VideoTrimmer } from './components/videoTrim';
-export * from './output';
+export { default as isElementBelowLimit } from './utils/isElementBelowLimit';
+export { default as isHexColorString } from './utils/isHexColorString';
 export { default as useCORSProxy } from './utils/useCORSProxy';
 export { default as getStoryPropsToSave } from './app/story/utils/getStoryPropsToSave';
 export { default as useRefreshPostEditURL } from './utils/useRefreshPostEditURL';
