+++ packages/story-editor/src/testUtils/index.js
@@ -16,6 +16,3 @@
 
 export { default as firePointerEvent } from './firePointerEvent';
 export { default as renderWithTheme } from './renderWithTheme';
-export * from './queryByAriaLabel';
-export * from './queryByAutoAdvanceAfter';
-export * from './queryById';
