+++ packages/testing-library/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@googleforcreators/testing-library",
+  "private": true,
+  "engines": {
+    "node": ">= 16",
+    "npm": "~7"
+  },
+  "exports": {
+    "default": "./src/index.js"
+  },
+  "main": "./src/index.js",
+  "type": "module",
+  "sideEffects": false,
+  "dependencies": {
+  },
+  "devDependencies": {
+  },
+  "peerDependencies": {
+    "@testing-library/react": "^12.1.2"
+  }
+}
