+++ packages/wp-story-editor/package.json
@@ -31,6 +31,7 @@
     "@googleforcreators/i18n": "*",
     "@googleforcreators/media": "*",
     "@googleforcreators/migration": "*",
+    "@googleforcreators/output": "*",
     "@googleforcreators/react": "*",
     "@googleforcreators/story-editor": "*",
     "@googleforcreators/tracking": "*",
